requirements.txt
llama-cpp-python
openai
llama-cpp-python
langchain-ollama
ollama